feature: moved templates to separate files
