+++ src/Documents/Sitemap.php
@@ -11,14 +11,17 @@ use Eightfold\XMLBuilder\Element;
 
 use Eightfold\HTMLBuilder\Element as HtmlElement;
 
-use JoshBruce\Site\FileSystemInterface;
+use JoshBruce\Site\HttpRequest;
 use JoshBruce\Site\File;
+use JoshBruce\Site\FileSystemInterface;
+
 use JoshBruce\Site\Content\Markdown;
 
 class Sitemap
 {
-    public static function create(FileSystemInterface $fileSystem): string
+    public static function create(File $file, HttpRequest $request): string
     {
+        $fileSystem = $file->fileSystem();
         $finder = self::finder($fileSystem);
 
         $files = self::files($finder, $fileSystem, true);
